+++ release-notes.md
@@ -4,6 +4,7 @@ RELEASE NOTES
 - use attributes when generate toString, in case no other important string is present
 - add support for css selector when setExcludeClasses
 - update webdriver version 2.49.0
+- add log at sendKeys method
 
 **Release Notes for Testy 2.0.0**
 - Change config default value for weblocator.log.suggestions=false